Make cable connections side aware
The mod Compact Machines adds a block that allows each of its faces
to be "connected" to a different block in a different dimension.
Because each face could be a different network/grid/system entirely
it requires the capabilities to be sided to be "proxied".

With this commit the path checks ask for the proper side of the
neighboring block instead of passing null to the capability checker
/getter.

This should not have any bad effects on the given cable system and
allows Integrated Dynamics to work with Compact Machines and other
Forge Capability proxying blocks (like Phantom Faces?) that require
a sided capability.