Fix items behaving weirdly when creating lists in the LP, Closes #446
This was caused by 39cdb353e67b7f1a05898e1f5233f4af4f2922cc which
moved player slot creation to happen after LP element slot creation.
But since LP list slots were only created client-side, this caused
a desync in slot id's on the server and client.
So now, a simple packet is sent from the server to the client
when a new list element is selected to resolve this desync.