Add safe-mode option, Closes #486
In cases where a network would become horribly broken because of a bug,
servers could get stuck in a repeated server crash loop.
This makes it possible to temporarily set the mod in safe-mode,
which means that networks won't tick anymore, values are not evaluated and no special rendering is done.
In this mode, the offending cables and parts can be removed.
When disabling this option again, everything will work again as normal.