Fix mech machines not working after chunk reload when network-powered
Closes #489

The problem was that non-ticking network elements were not properly re-validated (network set to non-null) after chunk reload, which caused them not to be able to find the energy network.

This patch is not applied to TileMultipartTicking yet, but if we would in the future add non-ticking elements there as well, we could reuse the helper functions introduced in this commit.