Fix proxies and delays not working anymore after being moved
This was a regression since the /setblock fix of #343.
This fix involves initializing the network twice.
Once before the TE data is set, and once after the TE data is set.
This is needed because there are cases (such as /setblock),
where the TE data is never set, but we still want to init the network
in those cases.
In all other cases, we potentially need data to init the network
after the TE data has been set, such as proxy and delay IDs.

Closes #626

Follow-up on b0d601b01f7d7a194351805de76ea2f9ea4b33b9