Fix ingredient extraction conflicting with index
In rare cases, an extraction could occur before the storage change
was properly indexed.

For example, if a crafter would produce an output,
and in the same tick, and exporter would export that crafted instance,
this instance would not be tracked in the index, and any
acknowledgements that this instance was ever produced would be lost,
while this is needed for the crafter.

This is now fixed by checking during extraction if the extracted
instance is in fact part of the index.
Otherwise, the extraction will fail.