Fix positioned parts being removed even if re-added in the same tick
This fixes an issue in IT where ingredient transfer would
stop because the index would become empty after a chunk reload.
This was because an interface removal would be removed via a delayed
list. IT sometimes re-adds parts in the same tick, but did not consider
that delayed list, which would cause the position (and index)
to be removed even if it was just re-added.
This fix simply makes sure that when a position is re-added,
it is removed from this delayed list.

Closes CyclopsMC/IntegratedTunnels#157