Ensure observer world reads happen in main world thread
This fixes many weird race conditions and mod conflicts
caused by reading the world within another thread in an
unsafe manner.

This change makes it so that the off-thread observer
only reads world data via a proxy, where this proxy
will update its view of the world every world tick,
so that the observer can access it in a thread-safe
manner.

Closes #1081